refactor: refactor code for review
