+++ src/components/bread-crumbs/bread-crumbs.module.css
@@ -1,0 +1,25 @@
+.linkContainer {
+  display: flex;
+  align-items: center;
+  gap: 18px;
+}
+
+.link {
+  text-decoration: none;
+  color: black;
+  position: relative;
+}
+
+.link:not(:last-child)::after {
+  content: '•';
+  position: absolute;
+  right: -12px;
+  top: 50%;
+  transform: translateY(-50%);
+  color: var(--clr-secondary-800);
+  font-size: 14px;
+}
+
+.activeLink {
+  color: var(--clr-accent-1);
+}
