+++ src/components/bread-crumbs/messages.ts
@@ -1,0 +1,4 @@
+export const messages = {
+  homeLink: 'Home',
+  notFoundLink: '404 Not Found',
+};
