+++ src/components/country-select/country-select.module.css
@@ -1,5 +1,5 @@
 .wrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
 }
