+++ src/components/header/bottom-header/bottom-header.module.css
@@ -1,8 +1,8 @@
 .bottomWrapper {
   display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	width: 1440px;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 1440px;
 }
 
 .bottomHeader {
