+++ src/components/header/top-header/top-header.module.css
@@ -1,15 +1,15 @@
 .topWrapper {
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-	width: 1440px;
-	padding: 0 10px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 1440px;
+  padding: 0 10px;
 }
 
 .topHeader {
-	display: flex;
-	align-items: center;
-	width: 100%;
-	height: 50px;
-	background-color: var(--clr-accent-2);
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: 50px;
+  background-color: var(--clr-accent-2);
 }
