+++ src/components/svg-builder/svg.module.css
@@ -18,4 +18,3 @@
   width: 40px;
   height: 30px;
 }
-
