+++ src/global.module.css
@@ -42,7 +42,7 @@ a:active {
   --clr-disabled-600: #5f5f5f;
 
   --clr-bg-1: #ffffff;
-  --clr-bg-2: #F6F5FF;
+  --clr-bg-2: #f6f5ff;
 
   --clr-line: #000000;
 }
