+++ src/pages/not-found-page/messages.ts
@@ -1,8 +1,5 @@
 export const messages = {
   errorMessage: '404 Not Found',
   oopsMessage: 'oops!',
-  homeLink: 'Home',
-  pagesLink: 'Pages',
-  notFoundLink: '404 Not Found',
   backButton: 'Back To Home',
 };
