+++ src/pages/register-page/register-page.module.css
@@ -1,39 +1,39 @@
 .container {
-    display: flex;
-    flex-direction: column;
-    padding: 2rem;
-    gap: 1rem;
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+  gap: 1rem;
 }
 
 .logoContainer {
-    display: flex;
-    gap: 1rem;
-    align-items: center;
-    justify-content: center;
+  display: flex;
+  gap: 1rem;
+  align-items: center;
+  justify-content: center;
 }
 
 .logo {
-    width: 50px;
-    height: 80px;
+  width: 50px;
+  height: 80px;
 }
 
 .signInHint {
-    text-align: center;
+  text-align: center;
 }
 
 .formContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
 }
 
 .formInput {
-    border: 1px solid var(--clr-primary-300);
-    border-radius: 4px;
-    padding: 0.5rem;
-    color: var(--clr-primary-500);
+  border: 1px solid var(--clr-primary-300);
+  border-radius: 4px;
+  padding: 0.5rem;
+  color: var(--clr-primary-500);
 }
 
 .formInput::placeholder {
-    color: var(--clr-primary-500);
+  color: var(--clr-primary-500);
 }
