+++ src/router/root.module.css
@@ -1,16 +1,17 @@
 .section {
-   display: flex;
-   justify-content: center;
-   flex-direction: column;
-   align-items: center;
-   width: 100%;
-   margin: 0 auto;
-   background-color: var(--clr-bg-2);
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  margin: 0 auto;
+  background-color: var(--clr-bg-1);
 
-   &.footer {
-    background-color: var(--clr-bg-1);
-   }
- }
+  &.footer {
+    background-color: var(--clr-bg-2);
+  }
+}
 
-
-
+.footer {
+  margin: 60px 0;
+}
