+++ src/router/root.tsx
@@ -1,5 +1,4 @@
 import { Outlet } from 'react-router-dom';
-import { Wrapper } from '../components/wrapper/wrapper';
 import { Header } from '../components/header/header';
 import { Footer } from '../components/footer/footer';
 import styles from './root.module.css';
@@ -9,14 +8,12 @@ export const Root = () => {
   return (
     <>
       <Header />
-      <Wrapper>
-        <main className={styles.section}>
-          <Outlet />
-        </main>
-        <footer className={classNames(styles.section, styles.footer)}>
-          <Footer />
-        </footer>
-      </Wrapper>
+      <main className={styles.section}>
+        <Outlet />
+      </main>
+      <footer className={classNames(styles.section, styles.footer)}>
+        <Footer />
+      </footer>
     </>
   );
 };
